[CW-7] starter for gateway and finish order

+++ server/order/internal/data/order.go
@@ -3,9 +3,9 @@ package data
 import "time"
 
 type Order struct {
-	ID         int       `json:"id"`
-	CustomerID int       `json:"customer_id"`
-	CarID      int       `json:"car_id"`
+	ID         int64     `json:"id"`
+	CustomerID int64     `json:"customer_id"`
+	CarID      int64     `json:"car_id"`
 	CreatedAt  time.Time `json:"created_at"`
 	UpdatedAt  time.Time `json:"updated_at"`
 }
